refactor(react): update stories and switch contentswitcher to class wrapper
